+++ src/pages/Contact.tsx
@@ -5,7 +5,7 @@ import React, {
   useState,
 } from 'react';
 import { makeStyles, Theme } from '@material-ui/core/styles';
-import { Trans, useTranslation } from 'react-i18next';
+import { useTranslation } from 'react-i18next';
 import {
   Button,
   CircularProgress,